backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
llama-index
requests
openai
pypdf
pymongo
pandas
